(PC-5475): Fix after rebase
