+++ src/pcapi/emails/beneficiary_soon_to_be_expired_bookings.py
@@ -5,18 +5,14 @@ from pcapi.models import Booking
 from pcapi.models import UserSQLEntity
 from pcapi.repository.feature_queries import feature_send_mail_to_users_enabled
 from pcapi.utils.mailing import DEV_EMAIL_ADDRESS
-from pcapi.utils.mailing import format_environment_for_email
 
 
 SUPPORT_EMAIL_ADDRESS = os.environ.get("SUPPORT_EMAIL_ADDRESS")
 
 
-# todo: FLOG redondance avec les notifs au JJ, voir avec fseguin qd son ticket sera mergé
 def build_soon_to_be_expired_bookings_recap_email_data_for_beneficiary(
     beneficiary: UserSQLEntity, bookings: typing.List[Booking]
 ) -> typing.Dict:
-    environment = format_environment_for_email()
-
     return {
         "FromEmail": SUPPORT_EMAIL_ADDRESS if feature_send_mail_to_users_enabled() else DEV_EMAIL_ADDRESS,
         "Mj-TemplateID": 1927224,
@@ -25,7 +21,6 @@ def build_soon_to_be_expired_bookings_recap_email_data_for_beneficiary(
         "Vars": {
             "user_firstName": beneficiary.firstName,
             "bookings": _extract_bookings_information_from_bookings_list(bookings),
-            "environment": environment,
         },
     }
