+++ src/pcapi/scripts/booking/notify_soon_to_be_expired_bookings.py
@@ -4,7 +4,9 @@ from operator import attrgetter
 
 from pcapi.core.bookings import conf
 import pcapi.core.bookings.repository as bookings_repository
+from pcapi.domain.user_emails import send_soon_to_be_expired_bookings_recap_email_to_beneficiary
 from pcapi.utils.logger import logger
+from pcapi.utils.mailing import send_raw_email
 
 
 def notify_soon_to_be_expired_bookings() -> None:
@@ -27,7 +29,8 @@ def notify_users_of_soon_to_be_expired_bookings(given_date: datetime.date = None
     notified_users = []
 
     if is_after_start_date:
-        for user, _bookings in expired_bookings_grouped_by_user.items():
+        for user, bookings in expired_bookings_grouped_by_user.items():
+            send_soon_to_be_expired_bookings_recap_email_to_beneficiary(user, bookings, send_raw_email)
             notified_users.append(user)
 
         logger.info(
